pyproject.toml
[tool.poetry]
name = "repl_python3_newcastle-3"
version = "0.1.0"
description = ""
authors = ["Kermitthefrog37 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
aiofiles = "^0.7.0"
discord = "^1.7.3"
utils = "^1.0.1"
apis = "^0.0.1"
better-profanity = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
